Added system to detect if player is close to an item
If player is close to an object with the "item" tag, a text appears above it. The text always faces the player and gets destroyed when the player gets further away from the object.
Created the "EmptyTextBoxObject" object. It gets cloned every time a new floating text needs to be created, so don't delete it.